my solution for task 2 and 3

+++ kernel/sysproc.c
@@ -41,7 +41,7 @@ sys_wait(void)
 uint64
 sys_sbrk(void)
 {
-  uint64 addr;
+  int addr;
   int n;
 
   if(argint(0, &n) < 0)
